+++ lyra/src/lib/__init__.py
@@ -1,13 +1,13 @@
 # pyright: reportUnusedImport=false
 from .extras import inj_glob, lgfmt
-from .errors import NotConnected
-from .utils import EmojiRefs, base_h, restricts_c
-from .music import cleanup
-from .lava import EventHandler
-from .lava.utils import (
-    access_equalizer,
+from .utils import LyraConfig, EmojiRefs, base_h, restricts_c
+from .connections import cleanup
+from .lava import (
+    EventHandler,
+    repeat_emojis,
     access_data,
     get_data,
-    repeat_emojis,
+    access_equalizer,
 )
 from .dataimpl import LyraDBClientType, LyraDBCollectionType, __init_mongo_client__
+from .errors import NotConnectedError
